refactor(stores): config imports
