+++ src/entities/payments-categories/index.ts
@@ -1,1 +1,10 @@
 export { usePaymentsCategories } from './model'
+export {
+    $categoriesStore,
+    $fetchPaymentCategoriesDate,
+    $servicesStore,
+    setupPaymentCategoriesRequestDate,
+    searchServices,
+    setupServices,
+    fetchPaymentCategoriesFx
+} from './model'
