+++ src/entities/payments-categories/model/index.ts
@@ -1,1 +1,10 @@
 export { usePaymentsCategories } from './use-payments-categories'
+export {
+    $categoriesStore,
+    $fetchPaymentCategoriesDate,
+    $servicesStore,
+    setupPaymentCategoriesRequestDate,
+    searchServices,
+    setupServices,
+    fetchPaymentCategoriesFx
+} from './store'
